+++ src/main/java/org/grouplens/reflens/RatingRecommender.java
@@ -24,5 +24,5 @@ import java.util.Map;
 import org.grouplens.reflens.data.ScoredObject;
 
 public interface RatingRecommender<U,I> {
-	public List<ScoredObject<I>> recommend(U user, Map<I,Float> ratings);
+	public List<ScoredObject<I>> recommend(U user, Map<I,Double> ratings);
 }
